requirements.txt
streamlit
yfinance
pandas
numpy
plotly
textblob
gspread
google-auth
lxml
beautifulsoup4
